VPAAMP-470: [Timeshift DAI] Confirm TimedMetadata Time relative to Stream
Reason for change:Added a EAP check and used the starttime of the periods
Test Procedure: Refer jira ticket VPAAMP-470
Priority: P2

+++ priv_aamp.cpp
@@ -8023,7 +8023,7 @@ void PrivateInstanceAAMP::ReportTimedMetadata(long long timeMilliseconds, const 
 
 		if (ISCONFIGSET_PRIV(eAAMPConfig_MetadataLogging))
 		{
-			AAMPLOG_WARN("aamp timedMetadata: [%ld] '%s'", (long)(timeMilliseconds), content.c_str());
+			AAMPLOG_WARN("aamp timedMetadata: [%lld] '%s'", timeMilliseconds, content.c_str());
 		}
 
 		if (!bSyncCall)
